pyproject.toml
[project]
name = "islamic-app-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.129.0",
    "motor>=3.7.1",
    "pwdlib[bcrypt]>=0.3.0",
    "pydantic-settings>=2.13.0",
    "pyjwt[crypto]>=2.11.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.40.0",
]
